# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESIT_Immo API",
        "description": "API REST pour la gestion locative — propriétaires, biens, locataires et contrats.",
        "contact": {
            "name": "Support RESIT_Immo",
            "email": "support@resit-immo.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/api/biens": {
            "get": {
                "tags": [
                    "Biens"
                ],
                "summary": "Lister tous les biens immobiliers",
                "operationId": "69042179cf04218810fe5cfa97308ba5",
                "responses": {
                    "200": {
                        "description": "Liste des biens avec leur propriétaire",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Bien"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Biens"
                ],
                "summary": "Créer un bien immobilier",
                "operationId": "c299fad90a3d8497f9162a3baf65daad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "type",
                                    "adresse",
                                    "prix",
                                    "proprietaire_id"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Appartement T3 Centre-ville"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "appartement",
                                        "enum": [
                                            "maison",
                                            "appartement"
                                        ]
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "5 Avenue Victor Hugo"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Lyon",
                                        "nullable": true
                                    },
                                    "prix": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 850
                                    },
                                    "proprietaire_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bien créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bien créé avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Bien"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/biens/{id}": {
            "get": {
                "tags": [
                    "Biens"
                ],
                "summary": "Afficher un bien avec son propriétaire et ses contrats",
                "operationId": "633204e8043aaa75b430caba272a9495",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du bien",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du bien",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Bien"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Bien introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Biens"
                ],
                "summary": "Mettre à jour un bien (remplacement complet)",
                "operationId": "56a011b611573aa3f61a60b90e2fedd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du bien",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Maison 4 pièces"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "maison",
                                            "appartement"
                                        ]
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "10 Rue du Moulin"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "prix": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "proprietaire_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bien mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bien mis à jour avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Bien"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Bien introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Biens"
                ],
                "summary": "Supprimer un bien",
                "operationId": "73cce0d4c874f4e72dd17d8f3c57ebc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du bien",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bien supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bien supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Bien introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Biens"
                ],
                "summary": "Mettre à jour un bien (mise à jour partielle)",
                "operationId": "6d57900da3f98cf7f97af3d233d85b0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du bien",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "prix": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bien mis à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Bien introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Bien": {
                "title": "Bien immobilier",
                "required": [
                    "titre",
                    "type",
                    "adresse",
                    "prix",
                    "proprietaire_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titre": {
                        "type": "string",
                        "example": "Appartement T3 Centre-ville"
                    },
                    "type": {
                        "type": "string",
                        "example": "appartement",
                        "enum": [
                            "maison",
                            "appartement"
                        ]
                    },
                    "adresse": {
                        "type": "string",
                        "example": "5 Avenue Victor Hugo"
                    },
                    "ville": {
                        "type": "string",
                        "example": "Lyon",
                        "nullable": true
                    },
                    "prix": {
                        "type": "number",
                        "format": "float",
                        "example": 850
                    },
                    "proprietaire_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Erreur de validation",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "nom": [
                                "Le nom est obligatoire."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum : Bearer {token}"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
...
}